descriptors to mod file; profiler to usb
fix tests
