+++ examples/filter_devices.rs
@@ -1,7 +1,7 @@
 /// This example shows how to use the USBFilter to filter out devices that match a certain criteria
 ///
 /// See [`USBFilter`] docs for more information
-use cyme::lsusb::profiler;
+use cyme::usb::profiler;
 use cyme::system_profiler::USBFilter;
 use cyme::usb::ClassCode;
